backend/requirements.txt
# FastAPI and Server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.30
aiosqlite>=0.20.0
greenlet>=3.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.2.0

# Validation
pydantic>=2.8.0
pydantic-settings>=2.3.0
email-validator>=2.2.0

# Google Gemini AI
google-generativeai>=0.7.0
llama-index-core
llama-index-llms-gemini
llama-index-embeddings-gemini
llama-index-vector-stores-postgres

# Offline Transcription
openai-whisper



# Utilities
python-dotenv>=1.0.1
aiofiles>=24.1.0

# Testing
pytest>=8.2.0
pytest-asyncio>=0.23.7
httpx>=0.27.0
